scripts/requirements.txt
adal==1.2.3
aiohttp==3.6.2
alabaster==0.7.12
antlr4-python3-runtime==4.7.2
applicationinsights==0.11.9
appnope==0.1.0
argcomplete==1.11.1
astroid==2.3.3
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.5
azure-batch==9.0.0
azure-cli==2.6.0
azure-cli-command-modules-nspkg==2.0.3
azure-cli-core==2.6.0
azure-cli-nspkg==3.0.4
azure-cli-telemetry==1.0.4
azure-common==1.1.25
azure-core==1.5.0
azure-cosmos==3.1.2
azure-datalake-store==0.0.48
azure-functions-devops-build==0.0.22
azure-graphrbac==0.60.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt-advisor==2.0.1
azure-mgmt-apimanagement==0.1.0
azure-mgmt-appconfiguration==0.4.0
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.52.0
azure-mgmt-batch==7.0.0
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-botservice==0.2.0
azure-mgmt-cdn==4.1.0rc1
azure-mgmt-cognitiveservices==5.0.0
azure-mgmt-compute==12.0.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==3.0.0rc13
azure-mgmt-containerservice==9.0.1
azure-mgmt-core==1.0.0
azure-mgmt-cosmosdb==0.14.0
azure-mgmt-datalake-analytics==0.2.1
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==0.1.0
azure-mgmt-deploymentmanager==0.2.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==2.2.0
azure-mgmt-eventhub==3.0.0
azure-mgmt-hdinsight==1.4.0
azure-mgmt-imagebuilder==0.2.1
azure-mgmt-iotcentral==3.0.0
azure-mgmt-iothub==0.12.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==0.5.0
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==0.2.0
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.2.1
azure-mgmt-media==2.1.0
azure-mgmt-monitor==0.9.0
azure-mgmt-msi==0.2.0
azure-mgmt-netapp==0.8.0
azure-mgmt-network==10.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.4.0
azure-mgmt-privatedns==0.1.0
azure-mgmt-rdbms==2.2.0
azure-mgmt-recoveryservices==0.4.0
azure-mgmt-recoveryservicesbackup==0.6.0
azure-mgmt-redhatopenshift==0.1.0
azure-mgmt-redis==7.0.0rc1
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.6.0
azure-mgmt-resource==9.0.0
azure-mgmt-search==2.1.0
azure-mgmt-security==0.1.0
azure-mgmt-servicebus==0.6.0
azure-mgmt-servicefabric==0.4.0
azure-mgmt-signalr==0.3.0
azure-mgmt-sql==0.18.0
azure-mgmt-sqlvirtualmachine==0.5.0
azure-mgmt-storage==9.0.0
azure-mgmt-trafficmanager==0.51.0
azure-mgmt-web==0.46.0
azure-multiapi-storage==0.3.2
azure-nspkg==3.0.2
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
bleach==3.1.4
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
colorama==0.4.3
colorlover==0.3.0
colour==0.1.5
comet-git-pure==0.19.15
comet-ml==3.1.11
configobj==5.0.6
cryptography==2.9
cycler==0.10.0
dash==1.12.0
dash-bootstrap-components==0.10.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
dask==2.14.0
-e git+https://github.com/AndreCNF/data-utils.git@b567cc82e649264ce0aec292a85403b1d866aff2#egg=data_utils
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
distributed==2.14.0
docutils==0.15.2
entrypoints==0.3
everett==1.0.2
fabric==2.5.0
falcon==2.0.0
fastparquet==0.3.3
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.4.0
fsspec==0.6.3
future==0.18.2
geojson==2.5.0
gitdb==4.0.4
gitdb2==2.0.6
GitPython==3.1.1
google==2.0.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graphviz==0.11.1
grpcio==1.28.1
HeapDict==1.0.1
htmlmin==0.1.12
httplib2==0.17.2
hug==2.6.1
humanfriendly==8.2
idna==2.9
imagesize==1.2.0
invoke==1.4.1
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
javaproperties==0.5.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsmin==2.2.2
jsondiff==1.2.0
jsonschema==3.0.2
jupyter-client==6.1.2
jupyter-core==4.6.3
jupyter-lsp==0.7.0
jupytext==1.4.2
kiwisolver==1.2.0
knack==0.7.1
lazy-object-proxy==1.4.3
livereload==2.6.1
llvmlite==0.31.0
lunr==0.5.6
lxml==4.5.0
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mkdocs==1.1
mkdocs-material==4.6.3
mkdocs-minify-plugin==0.2.1
mock==4.0.2
-e git+https://github.com/AndreCNF/model-interpreter.git@a180fdaa63e93e2ba1b90df5bfa5927ab863863f#egg=model_interpreter
modin==0.7.2
more-itertools==8.2.0
mpld3==0.3
msal==1.0.0
msal-extensions==0.1.3
msgpack==0.6.2
msrest==0.6.13
msrestazure==0.6.3
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.5
netifaces==0.10.9
nltk==3.4.5
notebook==6.0.3
numba==0.48.0
numpy==1.18.2
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.6.2
pdoc3==0.7.5
pdocs==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pixiedust==1.1.18
pkginfo==1.5.0.1
plotly==4.8.0
pluggy==0.13.1
portalocker==1.7.0
portray==1.3.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
py-spy==0.3.3
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pymdown-extensions==6.3
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==3.10.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.8
pytz==2019.1
PyYAML==5.3.1
pyzmq==19.0.0
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
scp==0.13.2
Send2Trash==1.5.0
-e git+https://github.com/AndreCNF/shap.git@65f1c58e91ec6fdfd2f2688af10583e24416c629#egg=shap
six==1.14.0
sklearn==0.0
smmap==3.0.2
smmap2==2.0.5
snowballstemmer==2.0.0
sortedcontainers==2.1.0
soupsieve==1.9.5
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sshtunnel==0.1.5
tabulate==0.8.7
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
thrift==0.13.0
toml==0.10.0
toolz==0.10.0
torch==1.4.0
tornado==5.1.1
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
ujson==1.35
uritemplate==3.0.1
urllib3==1.25.8
vsts==0.1.25
vsts-cd-manager==1.0.2
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.0
xgboost==0.90
xmltodict==0.12.0
yarl==1.4.2
yaspin==0.15.0
zict==2.0.0
zope.interface==5.1.0
